fix improper indexing

+++ spikeylab/stim/types/stimuli_classes.py
@@ -231,7 +231,7 @@ class Vocalization(AbstractStimulusComponent):
         signal = ((wavdata/amp_scale)*amp)
 
         if self._risefall > 0:
-            rf_npts = self._risefall * fs
+            rf_npts = int(self._risefall * fs)
             wnd = hann(rf_npts*2) # cosine taper
             signal[:rf_npts] = signal[:rf_npts] * wnd[:rf_npts]
             signal[-rf_npts:] = signal[-rf_npts:] * wnd[rf_npts:]
